move test utils from tensorflow_io/python/experimental/benchmark to tests/test_atds_avro/utils
