+++ tests/test_atds_avro/utils/data_source.py
@@ -16,7 +16,7 @@
 
 import hashlib
 
-from tensorflow_io.python.experimental.benchmark.hash_util import int_to_bytes
+from tests.test_atds_avro.utils.hash_util import int_to_bytes
 
 
 class DataSource:
