+++ tests/test_atds_avro/utils/generator/sparse_tensor_generator.py
@@ -20,10 +20,10 @@ import hashlib
 import numpy as np
 import tensorflow as tf
 
-from tensorflow_io.python.experimental.benchmark.generator.generator_base import (
+from tests.test_atds_avro.utils.generator.generator_base import (
     Generator,
 )
-from tensorflow_io.python.experimental.benchmark.hash_util import int_to_bytes
+from tests.test_atds_avro.utils.hash_util import int_to_bytes
 
 
 class ValueDistribution(enum.Enum):
